vrcomm-line-bot/requirements.txt
flask==3.0.3
line-bot-sdk==3.13.0
anthropic==0.28.0
openpyxl==3.1.2
gunicorn==22.0.0
gspread==6.1.2
google-auth==2.29.0
